删除多余的Buff抽象类
