+++ src/main/java/team/h2syj/fish/buff/Buff.java
@@ -1,7 +1,5 @@
 package team.h2syj.fish.buff;
 
-import java.util.List;
-
 import team.h2syj.fish.core.Biological;
 import team.h2syj.fish.core.Effect;
 
@@ -52,31 +50,10 @@ public interface Buff extends Effect {
         public abstract void execute(Biological target);
     }
 
-    // 回合开始时触发增益
-    abstract class TurnBeforeBuff extends AbstractBuff {
-        public TurnBeforeBuff(int turn) {
+    abstract class BaseBuff extends AbstractBuff {
+        public BaseBuff(int turn) {
             super(turn);
         }
-
-        public abstract void execute(Biological target);
-    }
-
-    // 施放攻击前
-    abstract class AttackBeforeBuff extends AbstractBuff {
-        public AttackBeforeBuff(int turn) {
-            super(turn);
-        }
-
-        public abstract void execute(Biological attacker, List<Biological> target);
-    }
-
-    // 施放攻击后
-    abstract class AttackAfterBuff extends AbstractBuff {
-        public AttackAfterBuff(int turn) {
-            super(turn);
-        }
-
-        public abstract void execute(Biological attacker, List<Biological> target);
     }
 
 }
